requirements.txt
flask
flask-cors
cloudinary
requests
python-dotenv
flask_cors